backend/requirements.txt
livekit-agents
livekit-plugins-google
python-dotenv